refactor: extract content-script constants into lib/content-constants.ts
Move SELECTORS, URL param constants, source keys, timeout configs,
and event names from content.ts into a shared constants module.
These are used as dep injections across many lib/ factories.

content.ts: 1153→1111 lines (-42)
