+++ apps/browser-extension/src/content.ts
@@ -45,91 +45,49 @@ import {
   parseGuardFieldNames,
   applyCollectionGuards,
 } from "./lib/collection-guards";
+import {
+  SELECTORS,
+  AUTO_EXPORT_PARAM,
+  AUTO_SYNC_PARAM,
+  AUTO_LIMIT_PARAM,
+  AUTO_MAX_PAGES_PARAM,
+  AUTO_MIN_AGE_PARAM,
+  AUTO_MAX_AGE_PARAM,
+  AUTO_SEARCH_PARAM,
+  AUTO_LOCATION_PARAM,
+  AUTO_KEYWORD_MODE_PARAM,
+  SAMPLE_NAME_PARAM,
+  JOB5156_HOST,
+  SEEK_HOST_SUFFIX,
+  JOB5156_PROFILE_URL_PREFIX,
+  SOURCE_KEYS,
+  SEEK_PROFILE_TYPE,
+  KEYWORD_MODE_CONCAT,
+  KEYWORD_MODE_SPACED,
+  JOB51_PAGE_COOLDOWN_MS,
+  JOB51_RATE_LIMIT_ERROR_MESSAGE,
+  API_CAPTURE_SOURCE,
+  EXTERNAL_ACCESS_KEY,
+  PAGE_BRIDGE_REQUEST_EVENT,
+  PAGE_BRIDGE_RESPONSE_EVENT,
+  PAGE_BRIDGE_REQUEST_ATTR,
+  PAGE_BRIDGE_RESPONSE_ATTR,
+  JOB51_NEXT_PAGE_EVENT,
+  CONTENT_SCRIPT_SOURCE,
+  JOB5156_DETAIL_FETCH_TIMEOUT_MS,
+  JOB5156_DETAIL_FETCH_CONCURRENCY,
+  JOB51_DETAIL_FETCH_TIMEOUT_MS,
+  JOB51_DETAIL_FETCH_CONCURRENCY,
+  DEFAULT_SEEK_PAGE_SIZE,
+  LATEST_AUTO_SYNC_SUMMARIES_STORAGE_KEY,
+} from "./lib/content-constants";
 
 /**
  * 智通直聘 Resume Collector - Content Script
  * Extracts resume data from hr.job5156.com/search page
  */
 
-// CSS Selectors based on DOM analysis
-const SELECTORS = {
-  listContainer: ".el-checkbox-group.resume-search-item-list-content-block",
-  resumeCard: ".list-content__li_part",
-  name: ".item-title-part1 a.name, a.name",
-  activityStatus: ".date-type-diff-text-block",
-  basicInfoRow: ".basic-line",
-  basicInfoItem: ".basic-line__text",
-  locationItem: ".resume-search-item-search-addre__span",
-  locationFallbackItem: ".text-truncate.text-center",
-  selfIntro: ".basic-keywords",
-  topRow: ".list-content__li__up-block",
-  topRowText: ".up-block__look-text",
-  workHistory: ".work-block",
-  workItem: ".work-item",
-  pagination: ".el-pagination",
-  nextPageBtn: ".el-pagination .btn-next",
-  seekPagination: 'nav[aria-label="Pagination of results"]',
-  seekTalentSearchPagination: 'nav[aria-label="PAGINATION_OF_RESULTS"]',
-  searchInput: ".el-autocomplete input.el-input__inner",
-  searchButton: ".resume-search-item-search-input-block__input-button",
-  // 51job eHire selectors
-  job51SearchInput: ".talent_search_keywords_input input.el-input__inner",
-  job51SearchButton: "button.search_button",
-  // Area selector (location filter modal)
-  areaTrigger: ".resume-search-item-search-addre",
-  areaModal: ".area-selector-item-block",
-  areaProvinceBlock:
-    ".area-selector-item-block__content__down__blcok:first-child",
-  areaCityBlock: ".area-selector-item-block__content__down__blcok:nth-child(2)",
-  areaDistrictBlock:
-    ".area-selector-item-block__content__down__blcok:nth-child(3)",
-  areaItem: ".down__blcok__select",
-  areaDistrictItem: ".down__block__big-select__block",
-  areaConfirmBtn: ".area-selector-item-block__footer .button-block.blue",
-  areaCancelBtn: ".area-selector-item-block__footer .button-block:not(.blue)",
-  areaSelectedCount: ".content__up__number__select",
-};
-
-const AUTO_EXPORT_PARAM = "tr_auto_export";
-const AUTO_SYNC_PARAM = "tr_auto_sync";
-const AUTO_LIMIT_PARAM = "tr_limit";
-const AUTO_MAX_PAGES_PARAM = "tr_max_pages";
-const AUTO_MIN_AGE_PARAM = "tr_min_age";
-const AUTO_MAX_AGE_PARAM = "tr_max_age";
-const AUTO_SEARCH_PARAM = "keyword";
-const AUTO_LOCATION_PARAM = "location";
-const AUTO_KEYWORD_MODE_PARAM = "tr_kw_mode";
-const SAMPLE_NAME_PARAM = "tr_sample_name";
-const JOB5156_HOST = "hr.job5156.com";
-const SEEK_HOST_SUFFIX = ".employer.seek.com";
-const JOB5156_PROFILE_URL_PREFIX = `https://${JOB5156_HOST}/resume/view/`;
-const SOURCE_KEYS = {
-  JOB5156: "job5156",
-  JOB51: "51job",
-  SEEK: "seek",
-  UNKNOWN: "unknown",
-};
-const SEEK_PROFILE_TYPE = "seek";
-const KEYWORD_MODE_CONCAT = "concat";
-const KEYWORD_MODE_SPACED = "spaced";
-const JOB51_PAGE_COOLDOWN_MS = 8000;
-const JOB51_RATE_LIMIT_ERROR_MESSAGE =
-  "51job 已触发访问频率限制，请60分钟后再试";
 let autoExportTriggered = false;
-const API_CAPTURE_SOURCE = "tr-resume-api";
-const EXTERNAL_ACCESS_KEY = "__TR_RESUME_DATA__";
-const PAGE_BRIDGE_REQUEST_EVENT = "trResumeBridgeRequest";
-const PAGE_BRIDGE_RESPONSE_EVENT = "trResumeBridgeResponse";
-const PAGE_BRIDGE_REQUEST_ATTR = "data-tr-resume-bridge-request";
-const PAGE_BRIDGE_RESPONSE_ATTR = "data-tr-resume-bridge-response";
-const JOB51_NEXT_PAGE_EVENT = "trJob51NextPageRequest";
-const CONTENT_SCRIPT_SOURCE = "tr-resume-content-script";
-const JOB5156_DETAIL_FETCH_TIMEOUT_MS = 5000;
-const JOB5156_DETAIL_FETCH_CONCURRENCY = 5;
-const JOB51_DETAIL_FETCH_TIMEOUT_MS = 8000;
-const JOB51_DETAIL_FETCH_CONCURRENCY = 2;
-const DEFAULT_SEEK_PAGE_SIZE = 20;
-const LATEST_AUTO_SYNC_SUMMARIES_STORAGE_KEY = "latestAutoSyncSummaries";
 
 // Capture auto-sync params from the initial URL at module init (document_start).
 // SEEK's SPA may strip unknown query params via history.replaceState before the
